BACKEND/requirements.txt
fastapi[standard]
sqlalchemy
psycopg[binary]
python-dotenv
python-jose[cryptography]
bcrypt[argon2]
pytest 
httpx
alembic
sqlalchemy-data-model-visualizer
uuid6
graphviz
pydantic-settings